Moved back ApiProvider and useSessionState
