+++ frontend/src/api/useSessionState.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, describe, expect, test } from "vitest";
 
-import { ApiClient } from "@/api/ApiClient";
 import { WhoAmIRequestHandler } from "@/testing/api-mocks/RequestHandlers";
 import { overrideOnce, server } from "@/testing/server";
 import { renderHook, waitFor } from "@/testing/test-utils";
 
+import { ApiClient } from "./ApiClient";
 import useSessionState from "./useSessionState";
 
 describe("Test useSessionState", () => {
